Switch to Entity classes
Using a primitive as entity in Java is a fool's errand, since every lookup
the entity is used for (its primary purpose) requires (auto)boxing to
object. We also never calculate with entity ids.

int seems rather limiting if you think about it, especially when you would
want to encode information in the entity id, like the machine it's coming
from or whatever.

So we're using our own type-safe Entity class with long payload and
equals/hashcode impl.

+++ src/main/java/org/javesy/DefaultIdGenerator.java
@@ -1,6 +1,6 @@
 package org.javesy;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class DefaultIdGenerator
     implements EntityIdGenerator
@@ -8,11 +8,11 @@ public class DefaultIdGenerator
     /**
      * Entity id counter.
      */
-    private AtomicInteger entityCount = new AtomicInteger(0);
+    private AtomicLong entityCount = new AtomicLong(0);
 
     @Override
-    public int getNextEntityId()
+    public Long getNextEntityId()
     {
-        return entityCount.getAndIncrement();
+        return Long.valueOf( entityCount.getAndIncrement());
     }
 }
