+++ src/main/java/org/javesy/EntityIdGenerator.java
@@ -2,6 +2,6 @@ package org.javesy;
 
 public interface EntityIdGenerator
 {
-    int getNextEntityId();
+    Long getNextEntityId();
 }
